fixes regex
globalizing the regex the way it's being used currently to test the value was breaking text input. this effectively does the same as proposed, where any combination of hex characters, spaces, and newlines is valid to unlimited length as a single match group, but anything not those characters will be rejected by the field.